rename footer-links to fit naming convention
